+++ lib/cineworld_uk.rb
@@ -9,6 +9,7 @@ require_relative './cineworld_uk/internal/film_with_screenings_parser'
 require_relative './cineworld_uk/internal/name_parser'
 require_relative './cineworld_uk/internal/titleize'
 require_relative './cineworld_uk/internal/screening_parser'
+require_relative './cineworld_uk/internal/whatson_parser'
 require_relative './cineworld_uk/internal/website'
 
 require_relative './cineworld_uk/cinema'